% Generated by roxygen2 (4.0.0): do not edit by hand
\name{AddProp}
\alias{AddProp}
\title{Add a new property (species, pure phase, ...) to an input
buffer}
\usage{
AddProp(input, name, values, cat, kinpar = NULL, first = NULL)
}
\arguments{
\item{input}{The input buffer.}

\item{name}{Name of the new property to add (element, PURE
mineral,...)}

\item{values}{The value(s) of the property.}

\item{cat}{The category of the property. Must be one of
\code{c("tot","pphases","kin")}}

\item{kinpar}{Kinetics parameters for the KINETICS case.}

\item{first}{A block to be added in the first SOLUTION, as in
\code{RepSol} and \code{Distribute}.}
}
\value{
The new input buffer upon which Rphree can be called.
}
\description{
Function that adds a new property (species, pure phase, ...) to an
input buffer.
}
\details{
.. content for DETAILS ..
}
\author{
MDL
}

