% Generated by roxygen2 (4.0.0): do not edit by hand
\name{DistributeKin}
\alias{DistributeKin}
\title{DistributeKin}
\usage{
DistributeKin(input, prop, values, ident = "-m0")
}
\arguments{
\item{input}{The initial buffer. It must contain one single
SOLUTION.}

\item{prop}{The property (in this case, the mineral with KINETICS)}

\item{values}{The numeric value to distribute. Will be coerced to
char.}

\item{ident}{Identifier of the property (e.g., -m, -m0, -parm)}
}
\value{
A new input buffer upon which Rphree can be run.
}
\description{
Distribute kinetics parameters in an input buffer containing one
single SOLUTION and KINETIC block.
}
\author{
MDL
}

