% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DS_CO2sol}
\alias{DS_CO2sol}
\title{Duan and Sun model for solubility of CO2 in NaCl solutions}
\usage{
DS_CO2sol(Tk, Pmpa, mNaCl, verbose = FALSE)
}
\arguments{
\item{Tk}{Temperature in Kelvin}

\item{Pmpa}{Partial pressure of CO2 in MPa}

\item{mNaCl}{Molality (mol/kgw) of NaCl in the aqueous solution}

\item{verbose}{Logical. If TRUE, additional messages are
visualized.}
}
\value{
Numeric value of xCO2 (mass fraction) of CO2 in the
aqueous phase at equilibrium
}
\description{
Solubility of CO2 in NaCl solutions in mol/kg water Duan & Sun
(2003) Chemical Geology 193 (2003) 257-271. Adapted from excel
macro by B.M. Krooss, Aachen.
}
\details{
TODO
}
\author{
MDL
}

