% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Solub.R
\name{DS_SI_CO2}
\alias{DS_SI_CO2}
\title{Saturation Index for solubility of CO2(g) using Duan
equation of State}
\usage{
DS_SI_CO2(Tk, Pmpa, poy = TRUE, V = NULL, verbose = FALSE)
}
\arguments{
\item{Tk}{Temperature in Kelvin}

\item{Pmpa}{Pressure in MPa}

\item{poy}{Logical. If TRUE, the Poynting correction term is also
computed and applied to the result}

\item{V}{Optional average specific molar volume of CO2. If
unspecified, the standard value of 32.0E-3 l/mol is used}

\item{verbose}{Logical. If TRUE, edditional messages are given.}
}
\value{
The corrected SI to include in PHREEQC simulations of
CO2(g)
}
\description{
Calculate the Saturation Index SI of CO2(g) given T, P and using
the Duan equation of state. A Poynting correction factor is also
applied.
}
\author{
MDL
}

