% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Solub.R
\name{DS_ZCO2_var}
\alias{DS_ZCO2_var}
\title{Compressibility Z of CO2 by the Duan and Sun (2003) model}
\usage{
DS_ZCO2_var(T, Pmpa, verbose = FALSE)
}
\arguments{
\item{Pmpa}{Pressure in MPa}

\item{verbose}{Logical. If TRUE, additional messages are showed}

\item{Tk}{Temperature in Kelvin}
}
\value{
Numeric value of compressibility Z of CO2
}
\description{
Computes the virial compressibility Z of CO2 by Duan, alternate
solution method using the Anderson-Björck Method
}
\details{
NONE
}
\author{
MDL
}

