% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rphree_Solub.R
\name{PH2Ompa}
\alias{PH2Ompa}
\title{Vapor pressure of water}
\usage{
PH2Ompa(Tk)
}
\arguments{
\item{Tk}{Temperature in K}
}
\value{
Vapor pressure of water in MPa
}
\description{
Duan & Sun model to calculate vapor pressure of water
}
\author{
MDL
}
