% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Viz.R
\name{PlotMinDelta}
\alias{PlotMinDelta}
\title{Barplot of mineral deltas of an Rphree solution}
\usage{
PlotMinDelta(sol, lim = 1e-04, names = TRUE, line = TRUE)
}
\arguments{
\item{sol}{The Rphree solution to visualize}

\item{lim}{Bound for the mindeltas to plot (-lim < mindelta < lim)}

\item{names}{Logical (if TRUE, all phases) or character vector of
the phases to plot}

\item{line}{Logical. If TRUE, a vertical line discriminating
dissolved from precipitated mineral phases gets plotted}
}
\value{
the output of \code{\link{graphics::barplot}}
}
\description{
Plots mineral deltas of an Rphree solution
}
\author{
MDL
}

