% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RepSol}
\alias{RepSol}
\title{Replicate a SOLUTION}
\usage{
RepSol(sol, n, first = NULL)
}
\arguments{
\item{sol}{The initial phreeqc buffer that needs to be replicated.}

\item{n}{Number of replicates to generate.}

\item{first}{A new block which needs to be inserted only in the
first solution (typically, PUNCH and alike).}
}
\value{
A buffer upon which Rphree can be called.
}
\description{
Replicates an input buffer containing only one SOLUTION, taking
care of SOLUTION/KINETICS/PURE identifiers. Eventually insert a
block (e.g: PUNCH) under the first SOLUTION.
}
\details{
Analogous to standard \code{rep} function.
}
\author{
MDL
}

