% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RGetPhases}
\alias{RGetPhases}
\title{Workhorse function for extraction of}
\usage{
RGetPhases(lin, which = NULL, delta = TRUE)
}
\arguments{
\item{lin}{The solution outputted by Rphree}

\item{which}{String containing the name of the specific pphase to
look for. If omitted, all pphases are returned.}

\item{delta}{Logical. If TRUE, delta min are returned, if FALSE,
total amount}
}
\value{
A data.frame containing the results of equilibrium phases
}
\description{
Workhorse function for extraction of results from a solution,
specifically equilibrium phases.
}
\author{
MDL
}

