% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rphree_Utils.R
\name{RInputFromList}
\alias{RInputFromList}
\title{Transform a (formatted) solution in a new input buffer.}
\usage{
RInputFromList(lin, rem, title = NULL, signif = 10L)
}
\arguments{
\item{lin}{Output list from a previous Rphree simulation}

\item{rem}{Optional string vector containing. If missing, pH, pe,
temp and water are reported by default.}

\item{title}{optional TITLE to appear in the new input buffer}

\item{signif}{integer specifying how many significant digits will
be written in the buffer}
}
\value{
An input buffer ready for further manipulation and/or a
new Rphree call.
}
\description{
Transform a (formatted) solution in a new input buffer.
}
\details{
Details TODO
}
\author{
MDL
}
