% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RKS_SI_CO2}
\alias{RKS_SI_CO2}
\title{Calculate the Saturation Index of CO2(g) by
Redlich-Kwong-Soave EOS}
\usage{
RKS_SI_CO2(Tk, Pmpa, poy = TRUE, V = NULL)
}
\arguments{
\item{Tk}{Temperature in Kelvin}

\item{Pmpa}{Pressure in MPa}

\item{poy}{Logical. If TRUE, a Poynting correction term is applied}

\item{V}{Numeric value of partial molar volume of aqueous CO2. If
left unspecified, 32 l/mol is assumed.}
}
\value{
The corrected SI to include in PHREEQC simulations of
CO2(g)
}
\description{
Calculate SI of CO2(g) given T, P and con using the
RedlichKwongSoave eos and applying a Poynting correction factor ..
}
\author{
MDL
}

