% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Main.R
\name{Rphree}
\alias{Rphree}
\title{Run online PHREEQC with the shipped 2.17 version}
\usage{
Rphree(input = stop("No input specified.\\n"), sel, write = FALSE,
  out = "Rphree", db = "phreeqc.dat", n, punch = 0, format = TRUE,
  verbose = FALSE)
}
\arguments{
\item{input}{The prepared input buffer, in form of a character
vector, one line per element}

\item{sel}{Selection of output blocks to be included in the
formatted list. If missing, everything is included in the output.
See also \code{RPhreeCheckSel} for more details.}

\item{write}{Logical. If TRUE, PHREEQC will write its usual output
on file on disk}

\item{out}{If \code{write=TRUE}, this indicates the name of the
output file to write to. The suffix ".Rout" will be added
automatically}

\item{db}{Character vector: the database. If \code{length(db)==1},
"db" is interpreted as filename on disk (relative path from the
working directory); or else, it is a buffer (character vector) as
returned by \code{db <- RPhreeFile(dbfile, is.db=TRUE)}}

\item{n}{Integer, optional parameter (deprecated, will be
suppressed in future versions) containing the number of
simulations included in the input buffer. If unspecified, some
heuristics will be performed to know how many \code{SOLUTIONS} are
present in the input buffer}

\item{punch}{number of parameters included in the PUNCH block. Not
used if no PUNCH statement is present in the input buffer}

\item{format}{Logical. If TRUE (which is default), the output
blocks are formatted as \code{data.frame}}

\item{verbose}{Logical. If TRUE, some additional checkpoint
message is outputted to the R console}
}
\value{
A list (or a list of lists) containing the output(s) of
the calculation
}
\description{
This function performs the actual PHREEQC calculations.
}
\details{
TODO
}
\examples{
\dontrun{ ## Load the database
pqcdb <- RPhreeFile(system.file("extdata", "phreeqc.dat", package="Rphree"), is.db=TRUE)
## Load a phreeqc script
ex1 <- RPhreeFile(system.file("extdata", "ex1.phrq", package="Rphree"), is.db=FALSE)
## Define a selection
mysel <- c(kin=FALSE, tot=TRUE, desc=TRUE, species = TRUE, pphases = TRUE, SI=TRUE, punch =FALSE)
## Run PHREEQC!
equipqc <- Rphree(ex1, db=pqcdb, sel=mysel, write=FALSE)
}
}
\author{
MDL
}

