% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RPhreeCheckSel}
\alias{RPhreeCheckSel}
\title{Generate and/or check the selection of output from Rphree}
\usage{
RPhreeCheckSel(sel)
}
\arguments{
\item{sel}{A named logical vector of length 7. If missing, the
function returns a valid selection vector with all components set
to FALSE.}
}
\value{
A well formed selection vector
}
\description{
Utility function to check if the selected output vector,
specifying the blocks included in the output of Rphree, is well
formed. The selection vector has the form: \code{ c(kin=FALSE,
tot=TRUE, desc=TRUE, species = TRUE, pphases = TRUE, SI=TRUE,
punch =FALSE)}, where the names of each vector component refer to
corresponding blocks in PHREEQC output.
}
\details{
This utility function is quite needed since \bold{the precise
order of this vector is in c-level hard-coded!}
}
\examples{
\dontrun{
## A valid example
mysel <- c(kin=FALSE, tot=TRUE, desc=TRUE, species = TRUE, pphases = TRUE, SI=TRUE, punch =FALSE)

## Let RPhreeCheckSel generate a selection:
sel <- RPhreeCheckSel()
sel <- !RPhreeCheckSel()
}
}
\author{
MDL
}

