% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RPhreeExt}
\alias{RPhreeExt}
\title{Workhorse function for extraction of informations from a
solution}
\usage{
RPhreeExt(sol, cat, prop)
}
\arguments{
\item{sol}{The solution outputted by Rphree}

\item{cat}{String: the category or output block where to look for
the property \code{prop}}

\item{prop}{The name of the property to look for}
}
\value{
Numeric of length 1, the inquired value for the property
}
\description{
Workhorse function for extraction of results from a solution,
primarily intended to be used by \code{\link{RPinfo}}, where it is
called inside a \code{sapply} statement.
}
\author{
MDL
}

