% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Utils.R
\name{RPhreeFile}
\alias{RPhreeFile}
\title{Read a normal PHREEQC input file and prepare it for Rphree}
\usage{
RPhreeFile(filein, is.db = FALSE, tabs = TRUE)
}
\arguments{
\item{filein}{PHREEQC input script to read.}

\item{is.db}{Logical. Set to TRUE if reading a database, for
performing additional checks.}

\item{tabs}{Logical, defaults to TRUE. Should we replace
tabulation with spaces?}
}
\value{
A buffer (character vector, one line per element)
containing a synthctically valid PHREEQC input
}
\description{
Reads a normal PHREEQC input file and prepares it for
\code{\link{Rphree}}
}
\details{
content for DETAILS: todo!
}
\examples{
\dontrun{inp <- RPhreeFile("test.phrq")
db <- RPhreeFile("db/phreeqc.dat",is.db=TRUE)}
}
\author{
MDL
}

