% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rphree_Utils.R
\name{RPinfo}
\alias{RPinfo}
\title{Extract specific values from an Rphree output}
\usage{
RPinfo(lin, cat = c("tot", "desc", "pphases", "master", "species", "kin",
  "SI"), prop, force = TRUE, flex = FALSE)
}
\arguments{
\item{lin}{The list containg the Rphree solution(s)}

\item{cat}{A string indicating the category (or output block) to
search in, must be one of: "tot, desc, pphases, master, species,
kin, SI"}

\item{prop}{The name of the inquired property (element, species,
mineral phase,\dots) whose value(s)}

\item{force}{Logical. If TRUE (default if left unspecified), a
valid numeric value (0) is returned even if the inquired property
is not found in the solution}

\item{flex}{Logical. If TRUE, do not rely on "ListInfo" in the
formatted solution list and perform heuristics to circumvent this
absence}
}
\value{
A numeric vector containing the inquired properties
}
\description{
This functions extracts informations from calculated Rphree
formatted solutions.
}
\examples{
\dontrun{
  RPinfo(SOL,"tot","Fe")
  RPinfo(SOL,"pphases","Anhydrite")
}
}
\author{
MDL
}
