% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RReadOut}
\alias{RReadOut}
\title{RReadOut}
\usage{
RReadOut(out)
}
\arguments{
\item{out}{The PHREEQC output file.}
}
\value{
An output list, as if the simulation would have being run
through
}
\description{
Reads a phreeqc output file and forms a results list as if the
calculation were made with Rphree.
}
\details{
.. content for DETAILS ..
}
\author{
MDL
}

