% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RReadOutKin}
\alias{RReadOutKin}
\title{RReadOutKin, import the output file of a kinetic simulation
into R}
\usage{
RReadOutKin(out, strip = TRUE, verbose = FALSE)
}
\arguments{
\item{out}{The PHREEQC output file.}

\item{strip}{don't know anymore}

\item{verbose}{logical. If TRUE more output is given to the
console (for debugging).}
}
\value{
An output list as per Rphree call.
}
\description{
Kinetics simulation: reads a phreeqc output file and forms an
output list - as if the calculation was made through
Rphree.
}
\author{
MDL
}

