% Generated by roxygen2 (4.0.0): do not edit by hand
\name{RReadPhases}
\alias{RReadPhases}
\title{extract the phases from a Rphree solution}
\usage{
RReadPhases(lin)
}
\arguments{
\item{lin}{The solution outputted by Rphree}
}
\value{
A data.frame containing the results of equilibrium phases
}
\description{
Workhorse function for extraction of results from a solution,
specifically equilibrium phases. Variant.
}
\author{
MDL
}

