% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SuppressSim}
\alias{SuppressSim}
\title{Suppress a simulation from an input buffer}
\usage{
SuppressSim(biginp, n = 1L)
}
\arguments{
\item{biginp}{The input buffer}

\item{n}{The number of the simulation to erase}
}
\value{
The new input buffer
}
\description{
Erases the SOLUTION number n (label, not position) from an input
buffer containing many.
}
\author{
MDL
}

