% Generated by roxygen2 (4.0.1): do not edit by hand
\name{VizAct}
\alias{VizAct}
\title{Plots activities or activity coefficients of a set of
solutions}
\usage{
VizAct(sol, spec, names, coef = TRUE, cutoff = 1e-12, ...)
}
\arguments{
\item{sol}{A list containing the output solutions. Must be of the
form \code{list(sol1,sol2)}}

\item{spec}{Char vector containing the species to be visualized}

\item{names}{Char vector of the names of the solutions}

\item{coef}{Logical. If TRUE, activity coefficients are
visualized, if FALSE activities.}

\item{cutoff}{Numeric. Lower bound for visualizing an activity}

\item{...}{Further optional parameters passed to
\code{\link{graphics::barplot}}}
}
\value{
Optional (invisible). A data.frame containing the
activities/act.coefficients (rows) for the given solutions
(columns)
}
\description{
Graphically compares - through barplots - the activity
coefficients for given species of a set of Rphree solutions. The input
}
\author{
MDL
}

