% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Viz.R
\name{VizMinDelta}
\alias{VizMinDelta}
\title{Plots mineral deltas of a set of solutions}
\usage{
VizMinDelta(sol, phases = NULL, names, ...)
}
\arguments{
\item{sol}{A list containing the output solutions. Must be of the
form \code{list(sol1,sol2)}}

\item{names}{Char vector of the names of the solutions}

\item{...}{Further optional parameters passed to
\code{\link{graphics::barplot}}}
}
\value{
Optional (invisible). A data.frame containing the
activities/act.coefficients (rows) for the given solutions
(columns)
}
\description{
Graphically compares - through barplots - the mineral deltas of a
set of Rphree solutions
}
\author{
MDL
}

