% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Rphree}
\alias{Rphree}
\title{Run PHREEQC}
\usage{
Rphree(input = stop("No input specified.\\n"), sel, write = FALSE,
  out = "Rphree", db = "phreeqc.dat", n, punch = 0, format = TRUE,
  verbose = FALSE)
}
\arguments{
\item{input}{The input buffer}

\item{sel}{Selection of output blocks to be included in the
formatted list. If missing, everything is included in the output}

\item{write}{Logical. If TRUE, PHREEQC will write its usual output
on file}

\item{out}{If \code{write=TRUE}, the name of the output file to write to}

\item{db}{The database}

\item{n}{Optional integer parameter containing the number of
simulations included in the input buffer}

\item{punch}{number of parameters included in the PUNCH block}

\item{format}{Logical. If TRUE, the output blocks are nicely formatted}

\item{verbose}{Logical. If TRUE, some additional checkpoint
message is outputted to the console}
}
\value{
A list (or a list of lists) containing the output(s) of
the calculation
}
\description{
This function performs the actual PHREEQC calculations.
}
\details{
TODO
}
\author{
MDL
}

