% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RPhreeWriteInp}
\alias{RPhreeWriteInp}
\title{RPhreeWriteInp}
\usage{
RPhreeWriteInp(ofile, input)
}
\arguments{
\item{ofile}{designated file the input will be written to.}

\item{input}{the input buffer.}
}
\description{
Write an input buffer to a file on disk.
}
\details{
It really just opens a file and uses "writeLines"" for writing the
input to it.
}
\author{
MDL
}

