% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rphree_Utils.R
\name{RReadOut}
\alias{RReadOut}
\title{RReadOut}
\usage{
RReadOut(out)
}
\arguments{
\item{out}{The PHREEQC output file.}
}
\value{
An output list, as if the simulation would have being run
through Rphree (the same blocks and the same names are returned)
}
\description{
Reads a phreeqc output file and forms a results list as if the
calculation were made with Rphree.
}
\details{
Currently all blocks are read. NOTE: the "charge balance" and
"Adjusted to redox" annotations are not read for the time being.
}
\author{
MDL
}

