% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rphree_Utils.R
\name{RReadOutKin}
\alias{RReadOutKin}
\title{RReadOutKin, import the output file of a kinetic simulation
into R}
\usage{
RReadOutKin(out, strip = TRUE, verbose = FALSE)
}
\arguments{
\item{out}{The PHREEQC output file.}

\item{strip}{logical. If TRUE, the "ListInfo" element will be
appended to the list.}

\item{verbose}{logical. If TRUE more output is given to the
console (for debugging).}
}
\value{
An output list as per Rphree call.
}
\description{
Kinetics simulation: reads a phreeqc output file and forms an
output list - as if the calculation was made through
Rphree.
}
\author{
MDL
}
